+++ cardano-api/src/Cardano/Api/ProtocolParameters.hs
@@ -1,0 +1,299 @@
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE StandaloneDeriving #-}
+{-# LANGUAGE TypeFamilies #-}
+
+-- | Protocol parameters.
+--
+-- This covers Protocol parameter updates that can be embedded in transactions.
+--
+-- TODO: add protocol parameters in ledger state queries.
+--
+module Cardano.Api.ProtocolParameters (
+    UpdateProposal(..),
+    ProtocolParametersUpdate(..),
+    makeShelleyUpdateProposal,
+
+    -- * Data family instances
+    AsType(..)
+  ) where
+
+import           Prelude
+
+import           Numeric.Natural
+import           Data.ByteString (ByteString)
+import qualified Data.Map.Strict as Map
+
+import           Control.Monad
+
+import           Cardano.Slotting.Slot (EpochNo)
+import qualified Cardano.Crypto.Hash.Class as Crypto
+import           Ouroboros.Consensus.Shelley.Eras (StandardShelley)
+import           Shelley.Spec.Ledger.BaseTypes (maybeToStrictMaybe)
+import qualified Shelley.Spec.Ledger.BaseTypes as Shelley
+import qualified Shelley.Spec.Ledger.PParams as Shelley
+
+import           Cardano.Api.Address
+import           Cardano.Api.Hash
+import           Cardano.Api.HasTypeProxy
+import           Cardano.Api.KeysByron
+import           Cardano.Api.KeysShelley
+import           Cardano.Api.SerialiseCBOR
+import           Cardano.Api.SerialiseTextEnvelope
+import           Cardano.Api.StakePoolMetadata
+import           Cardano.Api.TxMetadata
+import           Cardano.Api.Value
+
+
+-- ----------------------------------------------------------------------------
+-- Protocol updates embedded in transactions
+--
+
+newtype UpdateProposal = UpdateProposal (Shelley.Update StandardShelley)
+    deriving stock (Eq, Show)
+    deriving newtype (ToCBOR, FromCBOR)
+    deriving anyclass SerialiseAsCBOR
+
+instance HasTypeProxy UpdateProposal where
+    data AsType UpdateProposal = AsUpdateProposal
+    proxyToAsType _ = AsUpdateProposal
+
+instance HasTextEnvelope UpdateProposal where
+    textEnvelopeType _ = "UpdateProposalShelley"
+
+data ProtocolParametersUpdate =
+     ProtocolParametersUpdate {
+
+       -- | Protocol version, major and minor. Updating the major version is
+       -- used to trigger hard forks.
+       --
+       protocolUpdateProtocolVersion :: Maybe (Natural, Natural),
+
+       -- | The decentralization parameter. This is fraction of slots that
+       -- belong to the BFT overlay schedule, rather than the Praos schedule.
+       -- So 1 means fully centralised, while 0 means fully decentralised.
+       --
+       -- This is the \"d\" parameter from the design document.
+       --
+       protocolUpdateDecentralization :: Maybe Rational,
+
+       -- | Extra entropy for the Praos per-epoch nonce.
+       --
+       -- This can be used to add extra entropy during the decentralisation
+       -- process. If the extra entropy can be demonstrated to be generated
+       -- randomly then this method can be used to show that the initial
+       -- federated operators did not subtly bias the initial schedule so that
+       -- they retain undue influence after decentralisation.
+       --
+       protocolUpdateExtraPraosEntropy :: Maybe (Maybe ByteString),
+
+       -- | The maximum permitted size of a block header.
+       --
+       -- This must be at least as big as the largest legitimate block headers
+       -- but should not be too much larger, to help prevent DoS attacks.
+       --
+       -- Caution: setting this to be smaller than legitimate block headers is
+       -- a sure way to brick the system!
+       --
+       protocolUpdateMaxBlockHeaderSize :: Maybe Natural,
+
+       -- | The maximum permitted size of the block body (that is, the block
+       -- payload, without the block header).
+       --
+       -- This should be picked with the Praos network delta security parameter
+       -- in mind. Making this too large can severely weaken the Praos
+       -- consensus properties.
+       --
+       -- Caution: setting this to be smaller than a transaction that can
+       -- change the protocol parameters is a sure way to brick the system!
+       --
+       protocolUpdateMaxBlockBodySize :: Maybe Natural,
+
+       -- | The maximum permitted size of a transaction.
+       --
+       -- Typically this should not be too high a fraction of the block size,
+       -- otherwise wastage from block fragmentation becomes a problem, and
+       -- the current implementation does not use any sophisticated box packing
+       -- algorithm.
+       --
+       protocolUpdateMaxTxSize :: Maybe Natural,
+
+       -- | The constant factor for the minimum fee calculation.
+       --
+       protocolUpdateTxFeeFixed :: Maybe Natural,
+
+       -- | The linear factor for the minimum fee calculation.
+       --
+       protocolUpdateTxFeePerByte :: Maybe Natural,
+
+       -- | The minimum permitted value for new UTxO entries, ie for
+       -- transaction outputs.
+       --
+       protocolUpdateMinUTxOValue :: Maybe Lovelace,
+
+       -- | The deposit required to register a stake address.
+       --
+       protocolUpdateStakeAddressDeposit :: Maybe Lovelace,
+
+       -- | The deposit required to register a stake pool.
+       --
+       protocolUpdateStakePoolDeposit :: Maybe Lovelace,
+
+       -- | The minimum value that stake pools are permitted to declare for
+       -- their cost parameter.
+       --
+       protocolUpdateMinPoolCost :: Maybe Lovelace,
+
+       -- | The maximum number of epochs into the future that stake pools
+       -- are permitted to schedule a retirement.
+       --
+       protocolUpdatePoolRetireMaxEpoch :: Maybe EpochNo,
+
+       -- | The equilibrium target number of stake pools.
+       --
+       -- This is the \"k\" incentives parameter from the design document.
+       --
+       protocolUpdateStakePoolTargetNum :: Maybe Natural,
+
+       -- | The influence of the pledge in stake pool rewards.
+       --
+       -- This is the \"a_0\" incentives parameter from the design document.
+       --
+       protocolUpdatePoolPledgeInfluence :: Maybe Rational,
+
+       -- | The monetary expansion rate. This determines the fraction of the
+       -- reserves that are added to the fee pot each epoch.
+       --
+       -- This is the \"rho\" incentives parameter from the design document.
+       --
+       protocolUpdateMonetaryExpansion :: Maybe Rational,
+
+       -- | The fraction of the fee pot each epoch that goes to the treasury.
+       --
+       -- This is the \"tau\" incentives parameter from the design document.
+       --
+       protocolUpdateTreasuryCut :: Maybe Rational
+    }
+  deriving (Eq, Show)
+
+instance Semigroup ProtocolParametersUpdate where
+    ppu1 <> ppu2 =
+      ProtocolParametersUpdate {
+        protocolUpdateProtocolVersion     = merge protocolUpdateProtocolVersion
+      , protocolUpdateDecentralization    = merge protocolUpdateDecentralization
+      , protocolUpdateExtraPraosEntropy   = merge protocolUpdateExtraPraosEntropy
+      , protocolUpdateMaxBlockHeaderSize  = merge protocolUpdateMaxBlockHeaderSize
+      , protocolUpdateMaxBlockBodySize    = merge protocolUpdateMaxBlockBodySize
+      , protocolUpdateMaxTxSize           = merge protocolUpdateMaxTxSize
+      , protocolUpdateTxFeeFixed          = merge protocolUpdateTxFeeFixed
+      , protocolUpdateTxFeePerByte        = merge protocolUpdateTxFeePerByte
+      , protocolUpdateMinUTxOValue        = merge protocolUpdateMinUTxOValue
+      , protocolUpdateStakeAddressDeposit = merge protocolUpdateStakeAddressDeposit
+      , protocolUpdateStakePoolDeposit    = merge protocolUpdateStakePoolDeposit
+      , protocolUpdateMinPoolCost         = merge protocolUpdateMinPoolCost
+      , protocolUpdatePoolRetireMaxEpoch  = merge protocolUpdatePoolRetireMaxEpoch
+      , protocolUpdateStakePoolTargetNum  = merge protocolUpdateStakePoolTargetNum
+      , protocolUpdatePoolPledgeInfluence = merge protocolUpdatePoolPledgeInfluence
+      , protocolUpdateMonetaryExpansion   = merge protocolUpdateMonetaryExpansion
+      , protocolUpdateTreasuryCut         = merge protocolUpdateTreasuryCut
+      }
+      where
+        -- prefer the right hand side:
+        merge :: (ProtocolParametersUpdate -> Maybe a) -> Maybe a
+        merge f = f ppu2 `mplus` f ppu1
+
+instance Monoid ProtocolParametersUpdate where
+    mempty =
+      ProtocolParametersUpdate {
+        protocolUpdateProtocolVersion     = Nothing
+      , protocolUpdateDecentralization    = Nothing
+      , protocolUpdateExtraPraosEntropy   = Nothing
+      , protocolUpdateMaxBlockHeaderSize  = Nothing
+      , protocolUpdateMaxBlockBodySize    = Nothing
+      , protocolUpdateMaxTxSize           = Nothing
+      , protocolUpdateTxFeeFixed          = Nothing
+      , protocolUpdateTxFeePerByte        = Nothing
+      , protocolUpdateMinUTxOValue        = Nothing
+      , protocolUpdateStakeAddressDeposit = Nothing
+      , protocolUpdateStakePoolDeposit    = Nothing
+      , protocolUpdateMinPoolCost         = Nothing
+      , protocolUpdatePoolRetireMaxEpoch  = Nothing
+      , protocolUpdateStakePoolTargetNum  = Nothing
+      , protocolUpdatePoolPledgeInfluence = Nothing
+      , protocolUpdateMonetaryExpansion   = Nothing
+      , protocolUpdateTreasuryCut         = Nothing
+      }
+
+makeShelleyUpdateProposal :: ProtocolParametersUpdate
+                          -> [Hash GenesisKey]
+                          -> EpochNo
+                          -> UpdateProposal
+makeShelleyUpdateProposal params genesisKeyHashes epochno =
+    --TODO decide how to handle parameter validation
+    let ppup = toShelleyPParamsUpdate params in
+    UpdateProposal $
+      Shelley.Update
+        (Shelley.ProposedPPUpdates
+           (Map.fromList
+              [ (kh, ppup) | GenesisKeyHash kh <- genesisKeyHashes ]))
+        epochno
+
+toShelleyPParamsUpdate :: ProtocolParametersUpdate
+                       -> Shelley.PParamsUpdate StandardShelley
+toShelleyPParamsUpdate
+    ProtocolParametersUpdate {
+      protocolUpdateProtocolVersion
+    , protocolUpdateDecentralization
+    , protocolUpdateExtraPraosEntropy
+    , protocolUpdateMaxBlockHeaderSize
+    , protocolUpdateMaxBlockBodySize
+    , protocolUpdateMaxTxSize
+    , protocolUpdateTxFeeFixed
+    , protocolUpdateTxFeePerByte
+    , protocolUpdateMinUTxOValue
+    , protocolUpdateStakeAddressDeposit
+    , protocolUpdateStakePoolDeposit
+    , protocolUpdateMinPoolCost
+    , protocolUpdatePoolRetireMaxEpoch
+    , protocolUpdateStakePoolTargetNum
+    , protocolUpdatePoolPledgeInfluence
+    , protocolUpdateMonetaryExpansion
+    , protocolUpdateTreasuryCut
+    } =
+    Shelley.PParams {
+      Shelley._minfeeA     = maybeToStrictMaybe protocolUpdateTxFeePerByte
+    , Shelley._minfeeB     = maybeToStrictMaybe protocolUpdateTxFeeFixed
+    , Shelley._maxBBSize   = maybeToStrictMaybe protocolUpdateMaxBlockBodySize
+    , Shelley._maxTxSize   = maybeToStrictMaybe protocolUpdateMaxTxSize
+    , Shelley._maxBHSize   = maybeToStrictMaybe protocolUpdateMaxBlockHeaderSize
+    , Shelley._keyDeposit  = toShelleyLovelace <$>
+                               maybeToStrictMaybe protocolUpdateStakeAddressDeposit
+    , Shelley._poolDeposit = toShelleyLovelace <$>
+                               maybeToStrictMaybe protocolUpdateStakePoolDeposit
+    , Shelley._eMax        = maybeToStrictMaybe protocolUpdatePoolRetireMaxEpoch
+    , Shelley._nOpt        = maybeToStrictMaybe protocolUpdateStakePoolTargetNum
+    , Shelley._a0          = maybeToStrictMaybe protocolUpdatePoolPledgeInfluence
+    , Shelley._rho         = Shelley.truncateUnitInterval . fromRational <$>
+                               maybeToStrictMaybe protocolUpdateMonetaryExpansion
+    , Shelley._tau         = Shelley.truncateUnitInterval . fromRational <$>
+                               maybeToStrictMaybe protocolUpdateTreasuryCut
+    , Shelley._d           = Shelley.truncateUnitInterval . fromRational <$>
+                               maybeToStrictMaybe protocolUpdateDecentralization
+    , Shelley._extraEntropy    = mkNonce <$>
+                                   maybeToStrictMaybe protocolUpdateExtraPraosEntropy
+    , Shelley._protocolVersion = uncurry Shelley.ProtVer <$>
+                                   maybeToStrictMaybe protocolUpdateProtocolVersion
+    , Shelley._minUTxOValue    = toShelleyLovelace <$>
+                                   maybeToStrictMaybe protocolUpdateMinUTxOValue
+    , Shelley._minPoolCost     = toShelleyLovelace <$>
+                                   maybeToStrictMaybe protocolUpdateMinPoolCost
+    }
+  where
+    mkNonce Nothing   = Shelley.NeutralNonce
+    mkNonce (Just bs) = Shelley.Nonce
+                      . Crypto.castHash
+                      . Crypto.hashWith id
+                      $ bs
+
