+++ cardano-api/src/Cardano/Api/Typed.hs
@@ -359,7 +359,6 @@ module Cardano.Api.Typed (
     EpochNo(..),
     NetworkMagic(..),
     makeShelleyUpdateProposal,
-    toShelleyPParamsUpdate,
 
     -- ** Conversions
     --TODO: arrange not to export these
@@ -418,8 +417,6 @@ import qualified Data.Sequence.Strict as Seq
 import qualified Data.Set as Set
 import qualified Data.Vector as Vector
 
-import           Control.Monad
---import Control.Monad.IO.Class
 import           Control.Concurrent.STM
 import           Control.Tracer (nullTracer)
 
@@ -541,6 +538,7 @@ import           Cardano.Api.KeysByron
 import           Cardano.Api.KeysPraos
 import           Cardano.Api.KeysShelley
 import           Cardano.Api.NetworkId
+import           Cardano.Api.ProtocolParameters
 import           Cardano.Api.Script
 import           Cardano.Api.SerialiseBech32
 import           Cardano.Api.SerialiseCBOR
@@ -659,10 +657,6 @@ toShelleyTxOut :: TxOut Shelley -> Shelley.TxOut StandardShelley
 toShelleyTxOut (TxOut addr (TxOutAdaOnly _ value)) =
     Shelley.TxOut (toShelleyAddr addr) (toShelleyLovelace value)
 toShelleyTxOut (TxOut _addr (TxOutValue evidence _)) = case evidence of {}
-
-toShelleyLovelace :: Lovelace -> Shelley.Coin
-toShelleyLovelace (Lovelace l) = Shelley.Coin l
---TODO: validate bounds
 
 
 -- ----------------------------------------------------------------------------
@@ -1567,256 +1561,6 @@ toShelleyPoolParams StakePoolParameters {
                      . Text.pack
                      $ URI.uriToString id uri ""
 
-
--- ----------------------------------------------------------------------------
--- Protocol updates embedded in transactions
---
-
-newtype UpdateProposal = UpdateProposal (Shelley.Update StandardShelley)
-    deriving stock (Eq, Show)
-    deriving newtype (ToCBOR, FromCBOR)
-    deriving anyclass SerialiseAsCBOR
-
-instance HasTypeProxy UpdateProposal where
-    data AsType UpdateProposal = AsUpdateProposal
-    proxyToAsType _ = AsUpdateProposal
-
-instance HasTextEnvelope UpdateProposal where
-    textEnvelopeType _ = "UpdateProposalShelley"
-
-data ProtocolParametersUpdate =
-     ProtocolParametersUpdate {
-
-       -- | Protocol version, major and minor. Updating the major version is
-       -- used to trigger hard forks.
-       --
-       protocolUpdateProtocolVersion :: Maybe (Natural, Natural),
-
-       -- | The decentralization parameter. This is fraction of slots that
-       -- belong to the BFT overlay schedule, rather than the Praos schedule.
-       -- So 1 means fully centralised, while 0 means fully decentralised.
-       --
-       -- This is the \"d\" parameter from the design document.
-       --
-       protocolUpdateDecentralization :: Maybe Rational,
-
-       -- | Extra entropy for the Praos per-epoch nonce.
-       --
-       -- This can be used to add extra entropy during the decentralisation
-       -- process. If the extra entropy can be demonstrated to be generated
-       -- randomly then this method can be used to show that the initial
-       -- federated operators did not subtly bias the initial schedule so that
-       -- they retain undue influence after decentralisation.
-       --
-       protocolUpdateExtraPraosEntropy :: Maybe (Maybe ByteString),
-
-       -- | The maximum permitted size of a block header.
-       --
-       -- This must be at least as big as the largest legitimate block headers
-       -- but should not be too much larger, to help prevent DoS attacks.
-       --
-       -- Caution: setting this to be smaller than legitimate block headers is
-       -- a sure way to brick the system!
-       --
-       protocolUpdateMaxBlockHeaderSize :: Maybe Natural,
-
-       -- | The maximum permitted size of the block body (that is, the block
-       -- payload, without the block header).
-       --
-       -- This should be picked with the Praos network delta security parameter
-       -- in mind. Making this too large can severely weaken the Praos
-       -- consensus properties.
-       --
-       -- Caution: setting this to be smaller than a transaction that can
-       -- change the protocol parameters is a sure way to brick the system!
-       --
-       protocolUpdateMaxBlockBodySize :: Maybe Natural,
-
-       -- | The maximum permitted size of a transaction.
-       --
-       -- Typically this should not be too high a fraction of the block size,
-       -- otherwise wastage from block fragmentation becomes a problem, and
-       -- the current implementation does not use any sophisticated box packing
-       -- algorithm.
-       --
-       protocolUpdateMaxTxSize :: Maybe Natural,
-
-       -- | The constant factor for the minimum fee calculation.
-       --
-       protocolUpdateTxFeeFixed :: Maybe Natural,
-
-       -- | The linear factor for the minimum fee calculation.
-       --
-       protocolUpdateTxFeePerByte :: Maybe Natural,
-
-       -- | The minimum permitted value for new UTxO entries, ie for
-       -- transaction outputs.
-       --
-       protocolUpdateMinUTxOValue :: Maybe Lovelace,
-
-       -- | The deposit required to register a stake address.
-       --
-       protocolUpdateStakeAddressDeposit :: Maybe Lovelace,
-
-       -- | The deposit required to register a stake pool.
-       --
-       protocolUpdateStakePoolDeposit :: Maybe Lovelace,
-
-       -- | The minimum value that stake pools are permitted to declare for
-       -- their cost parameter.
-       --
-       protocolUpdateMinPoolCost :: Maybe Lovelace,
-
-       -- | The maximum number of epochs into the future that stake pools
-       -- are permitted to schedule a retirement.
-       --
-       protocolUpdatePoolRetireMaxEpoch :: Maybe EpochNo,
-
-       -- | The equilibrium target number of stake pools.
-       --
-       -- This is the \"k\" incentives parameter from the design document.
-       --
-       protocolUpdateStakePoolTargetNum :: Maybe Natural,
-
-       -- | The influence of the pledge in stake pool rewards.
-       --
-       -- This is the \"a_0\" incentives parameter from the design document.
-       --
-       protocolUpdatePoolPledgeInfluence :: Maybe Rational,
-
-       -- | The monetary expansion rate. This determines the fraction of the
-       -- reserves that are added to the fee pot each epoch.
-       --
-       -- This is the \"rho\" incentives parameter from the design document.
-       --
-       protocolUpdateMonetaryExpansion :: Maybe Rational,
-
-       -- | The fraction of the fee pot each epoch that goes to the treasury.
-       --
-       -- This is the \"tau\" incentives parameter from the design document.
-       --
-       protocolUpdateTreasuryCut :: Maybe Rational
-    }
-  deriving (Eq, Show)
-
-instance Semigroup ProtocolParametersUpdate where
-    ppu1 <> ppu2 =
-      ProtocolParametersUpdate {
-        protocolUpdateProtocolVersion     = merge protocolUpdateProtocolVersion
-      , protocolUpdateDecentralization    = merge protocolUpdateDecentralization
-      , protocolUpdateExtraPraosEntropy   = merge protocolUpdateExtraPraosEntropy
-      , protocolUpdateMaxBlockHeaderSize  = merge protocolUpdateMaxBlockHeaderSize
-      , protocolUpdateMaxBlockBodySize    = merge protocolUpdateMaxBlockBodySize
-      , protocolUpdateMaxTxSize           = merge protocolUpdateMaxTxSize
-      , protocolUpdateTxFeeFixed          = merge protocolUpdateTxFeeFixed
-      , protocolUpdateTxFeePerByte        = merge protocolUpdateTxFeePerByte
-      , protocolUpdateMinUTxOValue        = merge protocolUpdateMinUTxOValue
-      , protocolUpdateStakeAddressDeposit = merge protocolUpdateStakeAddressDeposit
-      , protocolUpdateStakePoolDeposit    = merge protocolUpdateStakePoolDeposit
-      , protocolUpdateMinPoolCost         = merge protocolUpdateMinPoolCost
-      , protocolUpdatePoolRetireMaxEpoch  = merge protocolUpdatePoolRetireMaxEpoch
-      , protocolUpdateStakePoolTargetNum  = merge protocolUpdateStakePoolTargetNum
-      , protocolUpdatePoolPledgeInfluence = merge protocolUpdatePoolPledgeInfluence
-      , protocolUpdateMonetaryExpansion   = merge protocolUpdateMonetaryExpansion
-      , protocolUpdateTreasuryCut         = merge protocolUpdateTreasuryCut
-      }
-      where
-        -- prefer the right hand side:
-        merge :: (ProtocolParametersUpdate -> Maybe a) -> Maybe a
-        merge f = f ppu2 `mplus` f ppu1
-
-instance Monoid ProtocolParametersUpdate where
-    mempty =
-      ProtocolParametersUpdate {
-        protocolUpdateProtocolVersion     = Nothing
-      , protocolUpdateDecentralization    = Nothing
-      , protocolUpdateExtraPraosEntropy   = Nothing
-      , protocolUpdateMaxBlockHeaderSize  = Nothing
-      , protocolUpdateMaxBlockBodySize    = Nothing
-      , protocolUpdateMaxTxSize           = Nothing
-      , protocolUpdateTxFeeFixed          = Nothing
-      , protocolUpdateTxFeePerByte        = Nothing
-      , protocolUpdateMinUTxOValue        = Nothing
-      , protocolUpdateStakeAddressDeposit = Nothing
-      , protocolUpdateStakePoolDeposit    = Nothing
-      , protocolUpdateMinPoolCost         = Nothing
-      , protocolUpdatePoolRetireMaxEpoch  = Nothing
-      , protocolUpdateStakePoolTargetNum  = Nothing
-      , protocolUpdatePoolPledgeInfluence = Nothing
-      , protocolUpdateMonetaryExpansion   = Nothing
-      , protocolUpdateTreasuryCut         = Nothing
-      }
-
-makeShelleyUpdateProposal :: ProtocolParametersUpdate
-                          -> [Hash GenesisKey]
-                          -> EpochNo
-                          -> UpdateProposal
-makeShelleyUpdateProposal params genesisKeyHashes epochno =
-    --TODO decide how to handle parameter validation
-    let ppup = toShelleyPParamsUpdate params in
-    UpdateProposal $
-      Shelley.Update
-        (Shelley.ProposedPPUpdates
-           (Map.fromList
-              [ (kh, ppup) | GenesisKeyHash kh <- genesisKeyHashes ]))
-        epochno
-
-toShelleyPParamsUpdate :: ProtocolParametersUpdate
-                       -> Shelley.PParamsUpdate StandardShelley
-toShelleyPParamsUpdate
-    ProtocolParametersUpdate {
-      protocolUpdateProtocolVersion
-    , protocolUpdateDecentralization
-    , protocolUpdateExtraPraosEntropy
-    , protocolUpdateMaxBlockHeaderSize
-    , protocolUpdateMaxBlockBodySize
-    , protocolUpdateMaxTxSize
-    , protocolUpdateTxFeeFixed
-    , protocolUpdateTxFeePerByte
-    , protocolUpdateMinUTxOValue
-    , protocolUpdateStakeAddressDeposit
-    , protocolUpdateStakePoolDeposit
-    , protocolUpdateMinPoolCost
-    , protocolUpdatePoolRetireMaxEpoch
-    , protocolUpdateStakePoolTargetNum
-    , protocolUpdatePoolPledgeInfluence
-    , protocolUpdateMonetaryExpansion
-    , protocolUpdateTreasuryCut
-    } =
-    Shelley.PParams {
-      Shelley._minfeeA     = maybeToStrictMaybe protocolUpdateTxFeePerByte
-    , Shelley._minfeeB     = maybeToStrictMaybe protocolUpdateTxFeeFixed
-    , Shelley._maxBBSize   = maybeToStrictMaybe protocolUpdateMaxBlockBodySize
-    , Shelley._maxTxSize   = maybeToStrictMaybe protocolUpdateMaxTxSize
-    , Shelley._maxBHSize   = maybeToStrictMaybe protocolUpdateMaxBlockHeaderSize
-    , Shelley._keyDeposit  = toShelleyLovelace <$>
-                               maybeToStrictMaybe protocolUpdateStakeAddressDeposit
-    , Shelley._poolDeposit = toShelleyLovelace <$>
-                               maybeToStrictMaybe protocolUpdateStakePoolDeposit
-    , Shelley._eMax        = maybeToStrictMaybe protocolUpdatePoolRetireMaxEpoch
-    , Shelley._nOpt        = maybeToStrictMaybe protocolUpdateStakePoolTargetNum
-    , Shelley._a0          = maybeToStrictMaybe protocolUpdatePoolPledgeInfluence
-    , Shelley._rho         = Shelley.truncateUnitInterval . fromRational <$>
-                               maybeToStrictMaybe protocolUpdateMonetaryExpansion
-    , Shelley._tau         = Shelley.truncateUnitInterval . fromRational <$>
-                               maybeToStrictMaybe protocolUpdateTreasuryCut
-    , Shelley._d           = Shelley.truncateUnitInterval . fromRational <$>
-                               maybeToStrictMaybe protocolUpdateDecentralization
-    , Shelley._extraEntropy    = mkNonce <$>
-                                   maybeToStrictMaybe protocolUpdateExtraPraosEntropy
-    , Shelley._protocolVersion = uncurry Shelley.ProtVer <$>
-                                   maybeToStrictMaybe protocolUpdateProtocolVersion
-    , Shelley._minUTxOValue    = toShelleyLovelace <$>
-                                   maybeToStrictMaybe protocolUpdateMinUTxOValue
-    , Shelley._minPoolCost     = toShelleyLovelace <$>
-                                   maybeToStrictMaybe protocolUpdateMinPoolCost
-    }
-  where
-    mkNonce Nothing   = Shelley.NeutralNonce
-    mkNonce (Just bs) = Shelley.Nonce
-                      . Crypto.castHash
-                      . Crypto.hashWith id
-                      $ bs
 
 -- ----------------------------------------------------------------------------
 -- Operational certificates
